requirements.txt
pytest==7.4.4
pytest-html==4.1.1
pytest-bdd==7.0.1
playwright==1.40.0
pyyaml==6.0.1
pytest-playwright==0.4.3
allure-pytest==2.13.2
